save first (if document was modified) before running OpenMW

+++ apps/opencs/model/doc/runner.cpp
@@ -1,6 +1,7 @@
 
 #include "runner.hpp"
 
+#include "operation.hpp"
 
 CSMDoc::Runner::Runner() : mRunning (false)
 {
@@ -8,26 +9,36 @@ CSMDoc::Runner::Runner() : mRunning (false)
         this, SLOT (finished (int, QProcess::ExitStatus)));
 }
 
-void CSMDoc::Runner::start()
+void CSMDoc::Runner::start (bool delayed)
 {
-    QString path = "openmw";
+    if (!delayed)
+    {
+        QString path = "openmw";
 #ifdef Q_OS_WIN
-    path.append(QString(".exe"));
+        path.append(QString(".exe"));
 #elif defined(Q_OS_MAC)
-    QDir dir(QCoreApplication::applicationDirPath());
-    path = dir.absoluteFilePath(name);
+        QDir dir(QCoreApplication::applicationDirPath());
+        path = dir.absoluteFilePath(name);
 #else
-    path.prepend(QString("./"));
+        path.prepend(QString("./"));
 #endif
 
-    mProcess.start (path);
+        mProcess.start (path);
+    }
+
     mRunning = true;
     emit runStateChanged();
 }
 
 void CSMDoc::Runner::stop()
 {
-    mProcess.kill();
+    if (mProcess.state()==QProcess::NotRunning)
+    {
+        mRunning = false;
+        emit runStateChanged();
+    }
+    else
+        mProcess.kill();
 }
 
 bool CSMDoc::Runner::isRunning() const
@@ -39,4 +50,21 @@ void CSMDoc::Runner::finished (int exitCode, QProcess::ExitStatus exitStatus)
 {
     mRunning = false;
     emit runStateChanged();
+}
+
+
+CSMDoc::SaveWatcher::SaveWatcher (Runner *runner, Operation *operation)
+: QObject (runner), mRunner (runner)
+{
+    connect (operation, SIGNAL (done (int, bool)), this, SLOT (saveDone (int, bool)));
+}
+
+void CSMDoc::SaveWatcher::saveDone (int type, bool failed)
+{
+    if (failed)
+        mRunner->stop();
+    else
+        mRunner->start();
+
+    deleteLater();
 }
